CR_Finalized essay.md
I worked in VSCode and used an extension for word count. In the end, it told me I had 894 words, but when I put it into word and checked it was more. I assume it has to do with formatting, but I hope it is okay for you the way I did it. I will send you the screenshot of my VSCode word count in the e-mail.